pyproject.toml
[project]
name = "seam"
version = "0.1.0"
description = "A multi-tenant SaaS webhook integration hub for event-driven automation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Renato"},
]
keywords = ["webhook", "integration", "automation", "saas", "multi-tenant"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
    # Django
    "django>=5.1,<6.0",
    "django-ninja>=1.3,<2.0",
    "django-cors-headers>=4.4,<5.0",

    # Database
    "psycopg[binary]>=3.2,<4.0",

    # Async / Celery
    "celery[redis]>=5.4,<6.0",
    "django-celery-beat>=2.6,<3.0",
    "django-celery-results>=2.5,<3.0",

    # HTTP Client
    "httpx>=0.27,<1.0",

    # Security
    "cryptography>=43.0,<44.0",
    "pyjwt>=2.9,<3.0",

    # Server
    "gunicorn>=22.0,<23.0",

    # Spreadsheet export
    "openpyxl>=3.1,<4.0",

    # Utilities
    "python-decouple>=3.8,<4.0",
    "django-extensions>=3.2,<4.0",
    "email-validator>=2.1,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-django>=4.8,<5.0",
    "pytest-celery>=1.0,<2.0",
    "pytest-cov>=5.0,<6.0",
    "pytest-asyncio>=0.23,<1.0",
    "factory-boy>=3.3,<4.0",
    "httpx>=0.27,<1.0",
    "ruff>=0.6,<1.0",
    "mypy>=1.11,<2.0",
    "django-stubs>=5.0,<6.0",
    "pre-commit>=3.8,<4.0",
    "ipython>=8.27,<9.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

# ==========================================
# Tool Configurations
# ==========================================

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src"]
extend-exclude = ["*/migrations/*"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "S",   # flake8-bandit (security)
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "T20", # flake8-print
    "SIM", # flake8-simplify
]
ignore = ["S101", "S106", "S105", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "common", "config"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
strict = false
warn_unused_configs = true
warn_redundant_casts = true

[tool.django-stubs]
django_settings_module = "config.settings.development"
